# api/model/errors.go
package model

// import "golang.org/x/text/message"

// Error ...
// type Error struct {
	// Massage string `json:"message"`
// }

// StandardErrorModel ...
type StandardErrorModel struct {
	Error string  `json:"error"`
}